Combine adult and nestling captures
Ringing and recapture of adults is stored in the ringing table (Rengas), while ringing and capture of nestlings is stored in the nestling table (Pullit). In current code, when there are chicks known to be ringed, add a row for each known chick. This works for now, but we are missing data about multiple visits/captures of chicks during the nestling phase.