Deal with ring series that pass 100 (e.g. 99 - 00)
If we try and make a series using the last 2 digits of the ring number from 99:00 it will fill in all values 1 - 99. This leads to an inflation of expected chicks. Instead, we need to make a series from 99:100 and use this to construct the true ring number. We still need to get the last 2 digits after doing this, because this is how ring numbers or stored in the Nestling table.