Deal with chicks in Capture but not Nestling
Some chicks are expected from the Capture table (i.e. they are in the ring series from RingNumber - LastRingNr_Brood), but they have no record in the Nestling table. In these cases, they are still recorded but they have no CaptureDate/Time. We use the date/time information provided in the Capture data when this occurs.