Create stand alone func to calculate age
Based on discussion with Frank, Age_calc should just be based on the first year an individual was caught. This prevents issues where the first observed age is mistaken. Instead we just determine whether an individual was caught as a chick, in which case age is known, otherwise assume an individual was >2yo on first capture.