Add additional arg to calc_age for capturedate
There may be cases where an individual is captured multiple times in one year. We want to sort by both SampleYear and CaptureDate to ensure that we have the first capture of the first season. We don't use just CaptureDate to avoid issue with species that breed over multiple years (e.g. southern hemisphere).