Update CHO test to use R output
Using csv output for tests is inefficient because a) we need to save and reload every csv and b) we are testing the output AFTER READING FROM CSV FILE, rather than after the R output (which is a better test of the code). Therefore, we now test on the R output rather than the csv output.