Add VEL tests
Tests for expected values in VEL are added. N.B. There is still some unexpected behaviour with chicks that were measured at day 6 but not 13. Currently they are given a capture record for both, even in cases where there is no evidence for a capture on both days. This is not ideal, we only want to record a capture if there was proof the individual was measured in some way, otherwise we have no evidence it was captured.