Deal with cases of caught fledglings CHO
In CHO, there are a few cases where a bird is caught as a fledgling in the same year it was hatched. The 'Age' is given as 'first year', but it should be treated differently to a 'first year' bird caught in the next breeding season (i.e. EURING 3 v. EURING 4).