Add ExperimentID Y/N col in MON
In the future we will need to translate these experiments into our chosen categories, but for now we just add a Y/N category for whether an experiment occurred. We also only consider experiments that were listed in the brood data, we don't include any information that was associated with chicks or parents. This will be difficult until we understand how the cross-fostering information works.