Include BroodIDLaid/Fledged in MON
MON has a lot of cross-fostering. To include BroodIDLaid, we take the BroodID from the brood data where the individual was listed as a chick. For BroodIDFledged, we assume it's the same as BroodIDLaid unless a 'destination' brood was given. In this case, we give the BroodID of a nest that was active at that given location on the date of capture (from brood data table). We need to link to brood data because otherwise we don't known the ClutchTypeObserved, which is essential to defining BroodID.