Add check to test unique individual IDs in Individual_data
The output message highlights all the rows that are duplicates. So if a record is duplicated twice. Both the first row as well as the other two are highlighted in the message.
Rows to test this check are added in dummy_data.
Part of issue #43. Currently the check tests within populations. But what about among populations?