Add check to compare brood size and number of chicks captured
We expect that brood size is equal to number of chicks in Capture_data. If brood size is smaller than number of chicks, error. If brood size is larger than number of chicks, warning (because a chick could have died before ringing/measurement).
Add corresponding rows in dummy data and update documentation.

Currently this code uses Brood_data, Individual_data and Capture_data, but actually on the first two are necessary. See comment in #35.