Create whitelist data object
"whitelist" is a list consisting of 4 data frames (i.e., Brood_whitelist, Capture_whitelist, Individual_whitelist and Location_whitelist), each of which consist of 3 variables: PopID, unique record ID (i.e., BroodID, CaptureID, IndvID and LocationID), and CheckID. This whitelist object will be called in each single quality check function and anti-joined with the records that are flagged as warnings and errors.
The R data object "whitelist" will be available in the data folder of the package. It will be manually created here in the script whitelist.R, simply by adding rows.