Include whitelisting procedure in each individual check
Whitelisting procedure:
The whitelist R data object will be called in each single quality check function, by anti-joining it with the records that are flagged as warnings and errors. The result are records flagged as warnings and errors that are NOT whitelisted, which will be listed in the quality check report.
This procedure is now included in brood_check.R, individual_check.R and location_check.R. The procedure is not yet included in capture_check.R, because the single-column ID "CaptureID" has not been created yet.