Updates to assigning individual identifiers
Made changes to several files to better deal with assigning individual identifiers to (re-ringed) birds. make_IndIdentifier() replaces the previous make_ReRingTable(), and assigns an integer identifier to each bird (irrespective of whether they have been re-ringed or not). ALL birds are then assigned the number of the first ring ever given to them as an individual identifier.