Changes to age calculation
After this update, age is calculated from capture histories BEFORE subsetting the ringing data to contain only EDM sites. That way, age calculation can take into account individuals ringed in other areas (immigrants). Furthermore, minimum parent age in the brood data is now treated as "true" (not minimum) age if the bird is classified as a local recruit in the raw data (as discussed with Malcolm).