Updates to basic subsetting function
Tested the subsetting function and made several corrections so that it runs without issues. Included functionality to output either .csv or .RDS (equivalent to run_pipelines()). Added functionality to save a list of data into current R environment (needed for testing). Changed file paths to be compatible with Mac. Changed save path so that all output files are saved into a folder (marked by date) within the directory of the standard data.