Functionality for subsetting with conflicting species
Extended the subsetting function to deal with individuals of conflicting species via a function argument include_conflicting. When include_conflicting = TRUE, all individuals that were identified at least once as one of the species of interest are included in capture and individual data. If include_conflicting = FALSE, all individuals of conflicting species are removed from capture and individual data.