MacOS compatibility for extract_paradox_db()
When running on MacOS, the python function extract_paradox requires the path to be specified in full (i.e. without the "~" in the beginning, as returned from choose_directory()), so added functionality to re-write full pathname for use here. 

Also disabled the selection of python version unless operating system is windows, because: 
a) The path specified here is specific to Liam
b) reticulate::use_python() does not seem to change anything on my computer
c) The way I have set up the RStudio-reticulate-python environment, it seems to automatically select the correct version/environment (needs verification).