Added functionality to split captures of collapsed records
In Nest data, it can happen that two adults of the same sex are caught in the same nest and this is noted as "IndvID1/IndvID2". Pipeline now splits this into separate captures of IndvID1 and IndvID2 in Capture_data. (Note that in Brood_data, parent ID is set to NA in these cases since it is not clear who is the parent).