Add check C5: check order of ages
The age of subsequent captures of an individual should be correct. If the individual is caught as an adult before it was caught as a chick, an error will be flagged. Other cases where the age of the capture is larger than the age of the subsequent capture will be flagged as a warning.