:white_check_mark: Allow individual tests to run
Rename test files so that individual tests can be run using the `filter` argument in devtools::test_that(). For this to work, we no longer run all pipelines with `run_pipelines()` in setup, but instead use `format_XXX()` in each test tile. `run_pipelines()` is tested individually to make sure it still behaves as expected.