:bug: Fix age bug in UAN pipeline
Frank Adriaensen told us that the quality check report flagged some records where the breeding female was falsely assumed to be a chick in the nest.
It appeared that when determining the `RingAge` in the `create_individual_UAN()`, `Age_observed` was assumed to be in the UAN format, whereas it was already converted to the EURING format used in SPI-Birds.