:construction: Split warnings and errors in two reports
From now on, the quality check procedure puts emphasis on the potential errors because the warnings are quite sensitive to false negatives and may occasionally generate very long lists, which are not particularly helpful to data owner or user. The warnings and verified records will still be provided, but in a separate document.