:bug: Fix bug for chicks without known age in check C1a
In check C1a, chick mass values are checked per age class. When an age class has too few records, a message is added to the report that tells the owner/user that reliable reference values could not be made. This used to happen as well for chicks without known age (i.e., 'NA' as age class), but this message is no longer printed. The values of chicks without known age are still checked.