🩹 Update B6: use experimental information
Following feedback on the QR, we now use experiment information in check B6.

Warning if BroodSize < # number of chicks in Individual_data because chicks could have died before ringing. Warning if BroodSize > # number of chicks and brood is experimentally manipulated. Error if BroodSize > # number of chicks and brood was not experimentally manipulated. Close #171.