:fire: Remove check L2
Check L2 used to check whether all locations in Location_data were present in Brood_data or Capture_data, but that is too strict. Some boxes may be checked but have no broods, and location data is not species-specific, which becomes a problem when a subset of species are selected from a study site.