:pencil2: Add note on "unknown"/"unringed" individuals
Now applied to the HAR situation, but there is a general issue of how we deal with "unknown"/"unringed" individuals. In some pipelines, they are removed; in others they are kept. In v1.2 the added variables captureRingNumber/releaseRingNumber may help in standardizing the way we deal with these types of individuals.